Update patch set 2
Patch Set 2:

I've never encountered this issue on otus. I'll do some testing today to confirm whether it starts happening on otus after a while, but I haven't noticed any issues in my initial testing (and have never had issues in the past). So if I confirm this is a condor specific issue, it would make more sense to put this workaround in the condor tree rather than common.

What I'm more interested is why this is happening. When did this issue begin? Why did it? Have you tried adding logging to the relevant part of frameworks/base to see what the timestamps are? Is it stuck in 1970? Does it give a blip in the past then correct times? Does the timestamp start OK then go strange after a while? I haven't been hearing of this issue before, so I presume the relevant code to bring up the time stamp involving the kernel and HAL used to work.

Patch-set: 2