Update patch set 2
Patch Set 2:

@Sultan I have an m7, and the rotation didn't stop after some uptime, it didn't work at all.

The question WHY it stoped working is something I'm interested in, too. As it worked until the 5th March for sure. The next build I compiled was on the 12th I think, then it didn't work anymore.

I found this way somewhere in gerrit for another moto device and did that to my HTC One m7, too.

But why it stopped working, would be interesting

Patch-set: 2